Fixed CircleCI issues from PR#74 due to TypeError in connect.ts (#75)
* Reverted to CircleCI config from commit 3797441095b5da7bff2580d4614160bbbd84aeb1

The CircleCI crash was affiliated with a TypeError in the file connect.ts which is getting fixed in this PR too

Co-Authored-By: Rahul Marban <rahulmarban@gmail.com>

* Fixed TypeError overlap in connect.ts

If statement would always return true in connect.ts due to the variable "v" being typeof number and "version" being typeof string which had no overlap. So inside the if statement the variable "v" gets converted to string beforehand.

Co-authored-by: Rahul Marban <rahulmarban@gmail.com>