Fixed collector not working on calc.ts
When the component collector was created the type it collects does not exist in djs v14 `componentType: "BUTTON"` we have to use the new ones which use the ComponentType enum included with discord.js. https://discordjs.guide/popular-topics/collectors.html#basic-message-component-collector On testing without this change the collector would not even collect anything.